将 AppSettingCollection移动到 AppSettingCollection.cs 增加文件尾空行、AppSettingCollection类不应该是部分类
