+++ AppShortcutsModel/Properties/app.manifest
